# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduFlow API Documentation",
        "description": "Swagger OpenApi implementation for EduFlow",
        "contact": {
            "email": "support@eduflow.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/courses/{id}/enroll": {
            "post": {
                "tags": [
                    "Inscriptions"
                ],
                "summary": "S'inscrire à un cours (Paiement Stripe)",
                "operationId": "056e8b95bd8348a95a55404cc9b8db4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du cours",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "stripeToken"
                                ],
                                "properties": {
                                    "stripeToken": {
                                        "description": "Token généré par Stripe.js",
                                        "type": "string",
                                        "example": "tok_visa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inscription réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Inscription réussie !"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de paiement ou d'inscription"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-courses/{id}/students": {
            "get": {
                "tags": [
                    "Inscriptions"
                ],
                "summary": "Liste des étudiants inscrits à un cours (Enseignant uniquement)",
                "operationId": "80c0ecdb213b48ab081a71fb5e87d630",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des étudiants récupérée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "course_id": {
                                            "type": "integer"
                                        },
                                        "total_students": {
                                            "type": "integer"
                                        },
                                        "students": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-stats": {
            "get": {
                "tags": [
                    "Inscriptions"
                ],
                "summary": "Statistiques globales des revenus et inscriptions (Enseignant uniquement)",
                "operationId": "cd15211df8af935cfab62c5417b2a9cd",
                "responses": {
                    "200": {
                        "description": "Statistiques récupérées",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_revenue": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "total_enrollments": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{id}/withdraw": {
            "delete": {
                "tags": [
                    "Inscriptions"
                ],
                "summary": "Se désinscrire d'un cours",
                "operationId": "e02a70533c0228884562471bc4fea7cb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Désinscription réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Inscription introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/{courseId}/groups": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Récupérer la structure des groupes d'un cours",
                "description": "Affiche la répartition des étudiants par groupes (max 25 par groupe). Réservé aux enseignants.",
                "operationId": "32c638a42ef2db646f87fe7c75162155",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "L'ID du cours pour lequel on veut voir les groupes",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Structure des groupes récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "course_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "group_name": {
                                                        "type": "string",
                                                        "example": "Groupe A"
                                                    },
                                                    "student_count": {
                                                        "type": "integer",
                                                        "example": 25
                                                    },
                                                    "students": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès interdit ou erreur lors de la récupération",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Action non autorisée."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}